requirements.txt
progressbar2
gdown
gitpython
hickle
tensorboard
numpy
scikit-image
einops
torch
torchvision
opencv-python-headless
wandb 
pyyaml
boto3
transformers
python-dotenv
tqdm
